Minor changes to remove compiler warnings
Build/Test Cases Summary
Enabled Packages: Zoltan, Zoltan2
0) MPI_DEBUG => passed: passed=15,notpassed=0 (1.60 min)
1) SERIAL_RELEASE => passed: passed=9,notpassed=0 (1.05 min)
Other local commits for this build/test group: 47f74fa

+++ packages/zoltan/src/Utilities/DDirectory/DD_Set_Neighbor_Hash_Fn1.c
@@ -115,7 +115,8 @@ static unsigned int dd_nh1 (ZOLTAN_ID_PTR gid, int gid_length,
  unsigned int nproc, struct dd_nh1_struct* hashdata, ZOLTAN_HASH_FN *fn)
    {
    int id = (signed) *gid;
-   return (id < hashdata->max_gid) ? (id / hashdata->groupsize) : (id % nproc);
+   return (unsigned int) ((id < hashdata->max_gid) ? (id / hashdata->groupsize)
+                                                   : (id % nproc));
    }
 
 #ifdef __cplusplus
